[GITISSUE-112] refactor: restructure project layout and update paths
- Moved `pkg/` to `api/` to better reflect the purpose of the package as API-related code.
- Updated all import paths to reflect the new directory structure.
- Renamed `resources/` to `assets/` for consistency and clarity.
- Updated references to diagram images in `README.md` to point to the new `assets/` directory.
- Adjusted import paths in `cmd/main.go` and `tests/smoke/smoke_test.go` to match the new structure.
- No functional changes were made; this is purely a structural refactor.
