+++ api/handlers/anime_handlers_test.go
@@ -9,10 +9,10 @@ import (
 	"context"
 	"encoding/json"
 	"errors"
-	"myanimeapi/pkg/handlers"
-	"myanimeapi/pkg/middleware"
-	"myanimeapi/pkg/mocks"
-	"myanimeapi/pkg/models"
+	"myanimeapi/api/handlers"
+	"myanimeapi/api/middleware"
+	"myanimeapi/api/mocks"
+	"myanimeapi/api/models"
 	"net/http"
 	"net/http/httptest"
 	"testing"
