+++ api/routes/routes.go
@@ -7,8 +7,8 @@ package routes
 
 import (
 	"log"
-	"myanimeapi/pkg/handlers"
-	"myanimeapi/pkg/middleware"
+	"myanimeapi/api/handlers"
+	"myanimeapi/api/middleware"
 	"net/http"
 
 	_ "myanimeapi/cmd/docs" // This is required to load the docs generated by Swaggo
