+++ assets/architecture_diagram.png
