+++ assets/component_diagram.png
