+++ assets/sequence_diagram_for_anime_creation.png
