+++ assets/sequence_diagram_for_user_authentication.png
