+++ assets/sequence_diagram_for_user_registration.png
